+++ pkg/refactor/cmd/main.go
@@ -129,7 +129,7 @@ func initSLog() {
 		Level: lvl,
 	}))
 
-	if os.Getenv(util.DirektivLogJSON) == "console" {
+	if os.Getenv(util.DirektivLogFormat) == "console" {
 		slogger = slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{
 			Level: lvl,
 		}))
