+++ pkg/refactor/gateway/endpoints/context.go
@@ -1,4 +1,4 @@
-package gateway
+package endpoints
 
 // This file has been copied and modified from https://github.com/go-chi/chi
